[IMP] sale_amazon: add information about compliance with Amazon's security standards
In the case where a seller would make a request to get access to the Personally Indentifiable Information of its customers, Amazon may require the seller to fill out a second form about his data protection policy.
This commit makes it clear that compliance with Amazon's security standards is not guaranteed if the seller applying for Amazon MWS access keys is hosted on Odoo.

Additionnaly, a section is added in which the questions asked to Amazon sellers are answered regarding the Odoo Online and Odoo.sh platforms.